# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CastDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.CreateBookingRequest:
    type: object
  dto.CreateBookingResponse:
    properties:
      booking_id:
        example: 1024
        type: integer
    type: object
...
  dto.GenreDTO:
    properties:
      genre:
        type: string
      id:
        type: integer
    type: object
  dto.LocationDTO:
    properties:
      city:
        example: Jakarta
        type: string
      id:
        example: 1
        type: integer
    type: object
...
  dto.MovieDetailResponse:
    properties:
      casts:
        items:
          $ref: '#/definitions/dto.CastDTO'
        type: array
      category:
        example: 13+
        type: string
      directors:
        example: Christopher Nolan
        type: string
      duration:
        example: "2:49:00"
        type: string
      genres:
        items:
          $ref: '#/definitions/dto.GenreDTO'
        type: array
      id:
        example: 1
        type: integer
      poster:
        example: ' https://storage.tickitz.id/posters/interstellar_reborn.jpg'
        type: string
      release_date:
        example: "2025-03-15"
        type: string
      synopsis:
        example: Seorang astronot nekat melintasi lubang cacing demi menyelamatkan
          umat manusia dari kehancuran bumi yang semakin tak layak huni. Perjalanan
          melintasi ruang dan waktu membawa konsekuensi yang tak pernah ia bayangkan
          sebelumnya.
        type: string
      title:
        example: Interstellar Reborn
        type: string
    type: object
  dto.MoviePaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.MovieResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      status:
        type: string
    type: object
  dto.MovieResponse:
    properties:
      genres:
        items:
          $ref: '#/definitions/dto.GenreDTO'
        type: array
      id:
        type: integer
      poster:
        type: string
      release_date:
        type: string
      title:
        type: string
    type: object
...
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      next:
        type: string
      prev:
        type: string
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  dto.ResponseError:
    properties:
      error:
        example: internal server error/bad request
        type: string
      message:
        example: Failed get data/internal server error
        type: string
      status:
        example: error
        type: string
    type: object
...
  dto.SeatDTO:
    properties:
      row:
        example: C
        type: string
      seat_id:
        example: 145
        type: integer
      seat_number:
        example: 7
        type: integer
      seat_status:
        enum:
        - available
        - sold
        example: available
        type: string
      seat_type:
        example: Regular
        type: string
    required:
    - seat_status
    type: object
  dto.SeatPageResponse:
    properties:
      seats:
        items:
          $ref: '#/definitions/dto.SeatDTO'
        type: array
      summary:
        $ref: '#/definitions/dto.SummaryMovieDTO'
    type: object
  dto.ShowtimeDetailResponse:
    properties:
      cinema_id:
        type: integer
      cinema_logo:
        type: string
      cinema_name:
        type: string
      city:
        type: string
      date:
        type: string
      movie_poster:
        type: string
      price:
        type: integer
      showtime_id:
        type: integer
      time:
        type: string
    type: object
  dto.ShowtimeFilterRequest:
    properties:
      city:
        example: Jakarta
        type: string
      date:
        example: "2025-06-10T15:04:05Z"
        type: string
      show_time:
        example: hh:mm:ss
        type: string
    type: object
  dto.ShowtimeFilterResponse:
    properties:
      locations:
        items:
          $ref: '#/definitions/dto.LocationDTO'
        type: array
      showtime:
        items:
          $ref: '#/definitions/dto.ShowtimeItemDTO'
        type: array
    type: object
  dto.ShowtimeItemDTO:
    properties:
      cinema_id:
        example: 1
        type: integer
      cinema_logo:
        example: ' https://storage.tickitz.id/cinema/ebvid.png'
        type: string
      cinema_name:
        example: ebv.id Grand Indonesia
        type: string
      show_date:
        example: "2025-06-10"
        type: string
      show_time:
        example: "13:30:00"
        type: string
      showtime_id:
        example: 2
        type: integer
      ticket_price:
        example: 75000
        type: integer
    type: object
...
  dto.SummaryMovieDTO:
    properties:
      category:
        example: 13+
        type: string
      cinema_logo:
        example: https://storage.tickitz.id/cinema/cgv.png
        type: string
      cinema_name:
        example: CGV Grand Indonesia
        type: string
      movie_poster:
        example: https://storage.tickitz.id/posters/interstellar_reborn.jpg
        type: string
      movie_title:
        example: Interstellar Reborn
        type: string
      show_date:
        example: "2026-06-15T00:00:00Z"
        type: string
      show_time:
        example: "18:30"
        type: string
      ticket_price:
        example: 50000
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: limustadji@gmail.com
    name: alpha male
    url: https://github.com/L1mus
  description: This is a backend project for a web application called E-wallet.
  license:
    name: MIT
  title: Movie API
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      description: fetch all movie data to be displayed on the main page
      produces:
      - application/json
      responses:
        "200":
          description: Success to get data
          schema:
            $ref: '#/definitions/dto.MoviePaginationResponse'
        "500":
          description: Failed to get data
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Get All movies
      tags:
      - Movies
  /movies/:id:
    get:
      description: Get the Movie detail data
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieDetailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get Movie Detail
      tags:
      - Movies
  /movies/:id/showtime:
    get:
      consumes:
      - application/json
      description: Get the Showtime data filtered result
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload filter showtime
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ShowtimeFilterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ShowtimeFilterResponse'
        "400":
          description: bad request" "must be filled
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Get Showtime Filter
      tags:
      - Movies
  /movies/{id}/showtimes:
    get:
      description: Fetch schedule and cinema details by movie ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/dto.ShowtimeDetailResponse'
            type: array
      summary: Get Movie Showtime Details
      tags:
      - movie
  /order/booking:
    post:
      consumes:
      - application/json
      description: Create a new ticket reservation for selected cinema seats.
      parameters:
      - description: Payload containing showtime ID and selected seat IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Booking successfully created
          schema:
            $ref: '#/definitions/dto.CreateBookingResponse'
        "400":
          description: Bad Request - Invalid JSON request payload
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "406":
          description: Not Acceptable - Selected seats are already sold or unavailable
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Reserve Seats and Create Booking
      tags:
      - Orders
  /order/seats:
    get:
      consumes:
      - application/json
      description: Retrieve the seating plan and availability status (Available/Sold)
        for a specific showtime.
      parameters:
      - description: Showtime Identifier ID
        in: query
        name: showtime_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved seating layout data
          schema:
            $ref: '#/definitions/dto.SeatPageResponse'
        "400":
          description: Bad Request - Invalid query parameter
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized - Valid authentication token required
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get Cinema Seats Layout
      tags:
      - Orders
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token used for authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"